pyproject.toml
[project]
name = "parchmentpdf"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.1",
    "fpdf2>=2.8.7",
    "jinja2>=3.1.6",
    "python-multipart>=0.0.26",
    "uvicorn>=0.46.0",
]
